move v2 code to specific directories
