+++ pkg/asyncapi/v2/channel.go
@@ -1,4 +1,4 @@
-package asyncapi
+package asyncapiv2
 
 import (
 	"strings"
