+++ pkg/asyncapi/v2/extensions.go
@@ -1,4 +1,4 @@
-package asyncapi
+package asyncapiv2
 
 import (
 	"fmt"
