+++ pkg/asyncapi/v2/extensions_test.go
@@ -1,4 +1,4 @@
-package asyncapi
+package asyncapiv2
 
 import (
 	"sort"
