+++ pkg/asyncapi/v2/message.go
@@ -1,4 +1,4 @@
-package asyncapi
+package asyncapiv2
 
 import (
 	"strconv"
