+++ pkg/asyncapi/v2/message_test.go
@@ -1,4 +1,4 @@
-package asyncapi
+package asyncapiv2
 
 import (
 	"testing"
