+++ pkg/asyncapi/v2/schema.go
@@ -1,4 +1,4 @@
-package asyncapi
+package asyncapiv2
 
 import (
 	"github.com/lerenn/asyncapi-codegen/pkg/utils"
