+++ pkg/codegen/generators/templates/helpers.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"unicode"
 
-	"github.com/lerenn/asyncapi-codegen/pkg/asyncapi"
+	asyncapi "github.com/lerenn/asyncapi-codegen/pkg/asyncapi/v2"
 )
 
 // NamifyWithoutParams will convert a sentence to a golang conventional type name.
