+++ pkg/codegen/parse.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/asyncapi/parser-go/pkg/parser"
 	"github.com/ghodss/yaml"
-	"github.com/lerenn/asyncapi-codegen/pkg/asyncapi"
+	asyncapi "github.com/lerenn/asyncapi-codegen/pkg/asyncapi/v2"
 )
 
 // FromFile parses the AsyncAPI specification either from a YAML file or a JSON file.
