+++ test/v2/issues/101/asyncapi.gen.go
