+++ test/v2/issues/101/asyncapi.yaml
