+++ test/v2/issues/114/asyncapi.gen.go
