+++ test/v2/issues/49/asyncapi.yaml
