+++ test/v2/issues/73/asyncapi-1.0.0.yaml
