+++ test/v2/issues/73/asyncapi-2.0.0.yaml
