+++ test/v2/issues/73/v1/asyncapi.gen.go
