+++ test/v2/issues/73/v2/asyncapi.gen.go
