+++ test/v2/issues/74/asyncapi.gen.go
