+++ test/v2/issues/97/asyncapi.gen.go
