+++ test/v2/issues/99/asyncapi.gen.go
